Fix CVE-2020-1968 (Racoon Attack)
According to the OpenSSL security advisory:
https://www.openssl.org/news/secadv/20200909.txt
The vulnerability only affects static DH ciphersuites, and that the fix
implemented by the OpenSSL team (for private release) was to move these
ciphersuites to the weak-ssl-ciphers list.

Use OPENSSL_NO_WEAK_SSL_CIPHERS flag to move the affected "static" DH
ciphersuites out of the default build configuration. Perhaps in the same way
that the OpenSSL team did for their private release.